+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/update/MediaGroupUpdate.kt
@@ -1,0 +1,10 @@
+package com.github.insanusmokrassar.TelegramBotAPI.types.update
+
+import com.github.insanusmokrassar.TelegramBotAPI.types.UpdateIdentifier
+import com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts.MediaGroupMessage
+import com.github.insanusmokrassar.TelegramBotAPI.types.update.abstracts.BaseMessageUpdate
+
+data class MediaGroupUpdate(
+    override val updateId: UpdateIdentifier,
+    override val data: MediaGroupMessage
+) : BaseMessageUpdate